Refactor client-side architecture
As things stand now, here is how things are organized, starting from the
root directory of the Backbone application (/editorsnotes_app/src/)-

models/, collections/, and utils/ still live at the root. As do the
entry points for Browserify (index-base.js and index-admin.js). This
commit adds the folders 'admin/' and 'base/' to the root, which will
contain *all views*. Views that deal with __editiing items__ should go
in "admin/", while views that __do not__ should go in "base/". For now,
I've made each view an individual "package" (folder) within its
appropriate view folder. Templates for each view should go into a
"template" folder within that view/package.

This should clear things up a lot and separate out where pieces of code
should live.
