+++ editorsnotes_app/src/admin/document/templates/document.html
