+++ editorsnotes_app/src/admin/document/templates/zotero_item.html
