+++ editorsnotes_app/src/admin/document/templates/zotero_item_type_select.html
