+++ editorsnotes_app/src/admin/generic/templates/add_item_modal.html
