+++ editorsnotes_app/src/admin/generic/templates/add_or_select_item.html
