+++ editorsnotes_app/src/admin/generic/templates/confirm_delete.html
