+++ editorsnotes_app/src/admin/note/templates/note.html
