+++ editorsnotes_app/src/admin/note/templates/note_section_list.html
