+++ editorsnotes_app/src/admin/topic/topic_citations.js
@@ -1,52 +1,58 @@
 "use strict";
 
-var OrderedCollectionView = require('./generic/ordered_collection_base')
-  , $ = require('../jquery')
+var OrderedCollectionView = require('../generic/ordered_collection_base')
+  , $ = require('jquery')
+  , CitationSectionView = require('../note/note_section').citation
+  , CitationView
+
+CitationView = CitationSectionView.extend({
+  initialize: function () { this.render(); },
+  render: function () {
+    var template = require('../note/templates/note_section_citation.html')
+      , citation = this.model
+      , html
+
+    html = template({ns: {
+      'document': citation.get('document'),
+      'document_description': citation.get('document_description'),
+      'content': citation.get('notes')
+    }});
+
+    this.$el.html(html);
+    this.afterRender();
+  }
+});
 
 module.exports = OrderedCollectionView.extend({
+  itemViewConstructor: CitationView,
   render: function () {
     var that = this
-      , template = require('../templates/note_section_list.html');
+      , template = require('./templates/citation_list.html')
 
     this.$el.empty().html( template() );
-    this.$itemsEl = this.$('.note-section-list');
+    this.$itemsEl = this.$('.citation-list');
     this.renderItems();
+    this.$el.closest('body').addClass('editing');
     this.initSort({
       stop: function (event, ui) {
         $(this).removeClass('sort-active');
         if (ui.item.hasClass('add-section')) {
-          that.createSection(ui.item.data('section-type'), ui.item.index());
+          that.collection.add({}, { at: ui.item.index() });
           ui.item.remove();
         }
       }
     });
     this.initDrag();
     this.initStickyBar();
-
-    $('body').addClass('editing');
   },
-
   onAddItemView: function (view) {
     if (view.model.isNew()) {
       view.edit();
     }
   },
-  makeItemView: function (section) {
-    var SectionView = require('./edit_note_section')[section.get('section_type')];
-    return new SectionView({ model: section });
-  },
-  handleAddSectionButton: function (e) {
+  handleAddCitationButton: function (e) {
     e.preventDefault();
-    this.createSection( $(e.currentTarget).data('section-type') ); 
-  },
-  createSection: function (sectionType, idx) {
-    // Sort is false because sections are ordered by the index of their ID in
-    // the note's section_ordering field. Since this new section does not yet
-    // have an ID, it can't be sorted.
-    return this.collection.add(
-      { section_type: sectionType },
-      { at: idx }
-    );
+    this.collection.add({});
   },
 
   initStickyBar: function () {
@@ -55,8 +61,8 @@ module.exports = OrderedCollectionView.extend({
       , $bar = $barContainer.find('#citation-edit-bar')
 
     $window
-      .off('scroll.note-section-list')
-      .on('scroll.note-section-list', function () {
+      .off('scroll.citation-list')
+      .on('scroll.citation-list', function () {
         var scrollTop = $window.scrollTop()
           , offsetTop = $barContainer.offset().top
 
@@ -84,6 +90,10 @@ module.exports = OrderedCollectionView.extend({
           .html( $(this).html() )
           .css('width', that.$itemsEl.innerWidth() - 22)
       },
+      start: function () {
+        threshold = $('.citation-list').offset().top;
+        $(this).animate({ 'opacity': 0 }, 120);
+      },
       drag: function (e, ui) {
         ui.position.top = null;
 
@@ -93,11 +103,7 @@ module.exports = OrderedCollectionView.extend({
           ui.helper.show();
         }
       },
-      start: function () {
-        threshold = $('.note-section-list').offset().top;
-        $(this).animate({ 'opacity': 0 }, 120);
-      },
-      stop: function () {
+      stop: function () { 
         $(this).animate({ 'opacity': 1 }, 80);
       }
     });
