+++ editorsnotes_app/src/utils/templates/wysihtml5_toolbar.html
